requirements.txt
asgiref==3.2.10
Django==3.1
django-timezone-field==4.0
Faker==4.1.1
gunicorn==20.0.4
jsonify==0.5
python-dateutil==2.8.1
pytz==2020.1
random-username==1.0.2
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
djangorestframework